moved around stuff, cleaned up code
