+++ components/Calendar/Calendar.js
@@ -1,55 +1,14 @@
-// #FFFFFF - white
-// #F5F5F5 - off-white
-// #B0B4D4 - light purp
-// #84C7D0 - sea blue
-// #4BD2D2 - ugly blue
-
-
-import { useState, useEffect } from 'react';
-// import { Scheduler, WeekView, DayView, Appointments, Toolbar, DateNavigator, ViewSwitcher,}
-//import moment from 'moment';
-import { styled, alpha } from '@mui/material/styles';
-import Paper from '@mui/material/Paper';
-import { ViewState } from '@devexpress/dx-react-scheduler';
-import {
-  Scheduler,
-  WeekView,
-  Toolbar,
-  DateNavigator,
-  Appointments,
-  TodayButton,
-  AppointmentTooltip,
-} from '@devexpress/dx-react-scheduler-material-ui';
-import TooltipContent from './subcomponents/TooltipContent.js';
-import styles from '../../utils/styles/CalendarStyles/Calendar.module.css';
-
-
-const data = [
-  {
-    title: 'KEvin CLASS',
-    startDate: new Date(2022, 2, 21, 6, 35),
-    endDate: new Date(2022, 2, 21, 22, 30),
-
-    zoomlink: 'http://garbase.com'
-  },
-  {
-  title: 'Other Class',
-  startDate: new Date(2022, 2, 24, 6, 35),
-  endDate: new Date(2022, 2, 24, 10, 30),
-  zoomlink: 'http://assaffsaafs.com'
-},]
-
-
-const PREFIX = 'Pog';
+import { styled } from '@mui/material/styles';
+import { WeekView } from '@devexpress/dx-react-scheduler-material-ui';
 
 const classes = {
-  todayCell: `${PREFIX}-todayCell`,
-  dayCell: `${PREFIX}-dayCell`,
-  todayHeader: `${PREFIX}-today`,
-  dayHeader: `${PREFIX}-weekday`,
+  todayCell: `todayCell`,
+  dayCell: `dayCell`,
+  todayHeader: `today`,
+  dayHeader: `weekday`,
 };
 
-const StyledWeekViewTimeTableCell = styled(WeekView.TimeTableCell)(({ theme }) => ({
+export const StyledWeekViewTimeTableCell = styled(WeekView.TimeTableCell)(({ theme }) => ({
   [`&.${classes.todayCell}`]: {
     backgroundColor: 'white',
     borderRight: 'none',
@@ -72,7 +31,8 @@ const StyledWeekViewTimeTableCell = styled(WeekView.TimeTableCell)(({ theme }) =
   },
 }));
 
-const StyledWeekViewDayScaleCell = styled(WeekView.DayScaleCell)(({ theme }) => ({
+
+export const StyledWeekViewDayScaleCell = styled(WeekView.DayScaleCell)(({ theme }) => ({
   [`&.${classes.todayHeader}`]: {
     backgroundColor: 'white',
   },
@@ -81,7 +41,8 @@ const StyledWeekViewDayScaleCell = styled(WeekView.DayScaleCell)(({ theme }) => 
   },
 }));
 
-const TimeTableCell = (props) => {
+
+export const TimeTableCell = (props) => {
   const { startDate } = props;
   const date = new Date(startDate);
 
@@ -93,7 +54,8 @@ const TimeTableCell = (props) => {
   return <StyledWeekViewTimeTableCell {...props} />;
 };
 
-const DayScaleCell = (props) => {
+
+export const DayScaleCell = (props) => {
   const { startDate, today } = props;
   if (today) {
     return <StyledWeekViewDayScaleCell {...props} className={classes.todayHeader} />;
@@ -102,41 +64,3 @@ const DayScaleCell = (props) => {
   }
   return <StyledWeekViewDayScaleCell {...props} />;
 };
-
-export default function Calendar () {
-  const [appointmentData, setAppointmentData] = useState([]);
-
-
-  useEffect(() => {
-    setAppointmentData(data);
-  }, []);
-
-
-
-
-  return (
-    <div>
-        <Paper elevation={6} className={styles.paper}>
-        <Scheduler
-          data={appointmentData}
-          height={'800'}
-        >
-          <ViewState />
-          <WeekView
-            startDayHour={9}
-            endDayHour={22}
-            timeTableCellComponent={TimeTableCell}
-            dayScaleCellComponent={DayScaleCell}
-          />
-          <Toolbar />
-          <DateNavigator />
-          <TodayButton />
-          <Appointments />
-          <AppointmentTooltip
-          // contentComponent={Tooltipcontent}
-          />
-        </Scheduler>
-      </Paper>
-    </div>
-  );
-}
