+++ components/Calendar/data/dummyData.js
@@ -1,0 +1,16 @@
+export const data = [
+    {
+      title: 'KEvin CLASS',
+      startDate: new Date(2022, 2, 21, 6, 35),
+      endDate: new Date(2022, 2, 21, 22, 30),
+
+      zoomlink: 'http://garbase.com'
+    },
+    {
+    title: 'Other Class',
+    startDate: new Date(2022, 2, 24, 6, 35),
+    endDate: new Date(2022, 2, 24, 10, 30),
+    zoomlink: 'http://assaffsaafs.com'
+  },
+]
+
