+++ utils/styles/CalendarStyles/Calendar.module.css
@@ -1,5 +1,5 @@
 .paper {
-  margin: 10px;
+  margin-top: 5%;
   border-radius: 20px;
   overflow: hidden;
 }
