+++ utils/styles/globals.css
@@ -69,7 +69,7 @@ a:hover {
 }
 
 /*CALENDAR DAY OF MONTH */
-.css-1ydj9ep-MuiTableCell-root .Cell-highlightedText {
+.css-i2thrp-MuiTableCell-root .Cell-highlightedText {
   color: #4BD2D2;
   font-weight: bold;
 }
